server: moves server back to server.js from index.js (allows us to separate repos)
